+++ core/primitives-core/Cargo.toml
@@ -21,6 +21,7 @@ enum-map.workspace = true
 num-rational.workspace = true
 serde.workspace = true
 serde_repr.workspace = true
+serde_with.workspace = true
 sha2.workspace = true
 thiserror.workspace = true
 schemars = { workspace = true, optional = true }
