+++ core/primitives/src/lib.rs
@@ -4,6 +4,8 @@ pub use near_primitives_core::account;
 pub use near_primitives_core::apply;
 pub use near_primitives_core::borsh;
 pub use near_primitives_core::config;
+pub use near_primitives_core::deterministic_account_id;
+pub use near_primitives_core::global_contract;
 pub use near_primitives_core::hash;
 pub use near_primitives_core::num_rational;
 pub use near_primitives_core::serialize;
@@ -16,7 +18,6 @@ pub mod block_header;
 pub mod challenge;
 pub mod chunk_apply_stats;
 pub mod congestion_info;
-pub mod deterministic_account_id;
 pub mod epoch_block_info;
 pub mod epoch_info;
 pub mod epoch_manager;
