+++ core/primitives/src/utils.rs
@@ -1,6 +1,5 @@
 #[cfg(feature = "clock")]
 use crate::block::BlockHeader;
-use crate::deterministic_account_id::DeterministicAccountStateInit;
 use crate::hash::{CryptoHash, hash};
 use crate::stateless_validation::ChunkProductionKey;
 use crate::types::{NumSeats, NumShards, ShardId};
@@ -8,6 +7,7 @@ use chrono;
 use chrono::DateTime;
 use near_crypto::{ED25519PublicKey, Secp256K1PublicKey};
 use near_primitives_core::account::id::{AccountId, AccountType};
+use near_primitives_core::deterministic_account_id::DeterministicAccountStateInit;
 use near_primitives_core::types::BlockHeight;
 use serde;
 use std::cmp::max;
