+++ runtime/runtime/src/deterministic_account_id.rs
@@ -5,11 +5,11 @@ use crate::{ActionResult, ApplyState};
 use near_parameters::StorageUsageConfig;
 use near_primitives::account::{Account, AccountContract};
 use near_primitives::action::DeterministicStateInitAction;
-use near_primitives::deterministic_account_id::DeterministicAccountStateInit;
 use near_primitives::errors::{IntegerOverflowError, RuntimeError};
 use near_primitives::receipt::Receipt;
 use near_primitives::trie_key::TrieKey;
 use near_primitives::types::AccountId;
+use near_primitives_core::deterministic_account_id::DeterministicAccountStateInit;
 use near_store::{StorageError, TrieUpdate};
 use near_vm_runner::logic::ProtocolVersion;
